0.9: Se añade BrandService

+++ src/main/java/inditex/productpriceservice/entities/Brand.java
@@ -29,10 +29,10 @@ public class Brand {
 
 
     // Getters y Setters
-    public long getId() {
+    public int getId() {
         return id;
     }
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
     public String getBrandName() {
